rename ana1Tests and associated run scripts
